Add files via upload
This script creates the matrices to compute the coupling coefficients of the VK plate using the VK operator. The matrices are second-order accurate approximations to d/dx and d/dy, acting on a 2D grid (the plate). The matrices can be composed to get d^2/dx^2,  d^2/dy^2, d^2/dxdy.