first pass diagonalisation of biharmonic
There is a lot of recalculation of coefficients, but this is just down to how the Dxx coeff functions are organised. will start extracting tis stage to `biharmdiag.m` before re-organising. The final block of the D+(Ny) daiagonal, seems to be a problem, likely a typo. Interestingly there is a reasonable amount of rounding error between multiplying by `1/h^4` and dividing the biharmonic by `h^4` (i.e. `biharm = biharm/h^4`) in general I think it is safe to say that multiplication  is to be preferred.